requirements.txt
ansible
ansible-runner
PyYAML
requests
python-terraform
google-auth
pytest